+++ lib/coinbase/wallet.rb
@@ -367,6 +367,14 @@ module Coinbase
       default_address.claimable_balance(asset_id, mode: mode, options: options)
     end
 
+    # Enumerates the historical balances for a given asset belonging of default address.
+    # The result is an enumerator that lazily fetches from the server, and can be iterated over,
+    # converted to an array, etc...
+    # @return [Enumerable<Coinbase::HistoricalBalance>] Enumerator that returns historical_balance
+    def historical_balances(asset_id)
+      default_address.historical_balances(asset_id)
+    end
+
     # Exports the Wallet's data to a Data object.
     # @return [Coinbase::Wallet::Data] The Wallet data
     def export